LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/style-not-sharing-float-expected.txt


FAIL Check that style is not sharing in the case of Container Queries assert_equals: Second item container should be 100px wide expected "rgb(0, 255, 0)" but got "rgb(0, 0, 0)"
